+++ demos/chinook/src/main/java/is/codion/framework/demos/chinook/tutorial/ClientTutorial.java
@@ -146,7 +146,7 @@ public final class ClientTutorial {
       createForeignKeySearchField(Album.ARTIST_FK)
               .columns(15);
       createTextField(Album.TITLE)
-              .action(control(ControlCode.INSERT))
+              .action(control(EditControl.INSERT).get())
               .columns(15);
       setLayout(gridLayout(2, 1));
       addInputPanel(Album.ARTIST_FK);